requirements.txt
dotenv==0.9.9
mysql-connector-python==9.3.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
python-dotenv==1.1.0
